+++ source/StopWatch/Helpers/CrossPlatformHelpers.cs
