+++ source/StopWatch/Helpers/InvokeExtensions.cs
