+++ source/StopWatch/Helpers/JiraTimeHelpers.cs
