+++ source/StopWatch/Helpers/NativeMethods.cs
