+++ source/StopWatch/Jira/DTO/AvailableTransitions.cs
