+++ source/StopWatch/Jira/DTO/Filter.cs
