+++ source/StopWatch/Jira/DTO/Issue.cs
