+++ source/StopWatch/Jira/DTO/SearchResult.cs
