+++ source/StopWatch/Jira/DTO/Transition.cs
