+++ source/StopWatch/Model/WatchTimer.cs
