+++ source/StopWatch/Settings/PersistedIssue.cs
