+++ source/StopWatch/Settings/Settings.cs
