+++ source/StopWatch/StopWatch.csproj
@@ -47,28 +47,28 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AboutForm.cs">
+    <Compile Include="UI\AboutForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="AboutForm.Designer.cs">
+    <Compile Include="UI\AboutForm.Designer.cs">
       <DependentUpon>AboutForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="CrossPlatformHelpers.cs" />
-    <Compile Include="DPApi.cs" />
-    <Compile Include="EditTimeForm.cs">
+    <Compile Include="Helpers\CrossPlatformHelpers.cs" />
+    <Compile Include="Helpers\DPApi.cs" />
+    <Compile Include="UI\EditTimeForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="EditTimeForm.Designer.cs">
+    <Compile Include="UI\EditTimeForm.Designer.cs">
       <DependentUpon>EditTimeForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="InvokeExtensions.cs" />
-    <Compile Include="JiraApi\AvailableTransitions.cs" />
-    <Compile Include="JiraApi\Transition.cs" />
+    <Compile Include="Helpers\InvokeExtensions.cs" />
+    <Compile Include="Jira\DTO\AvailableTransitions.cs" />
+    <Compile Include="Jira\DTO\Transition.cs" />
     <Compile Include="Jira\IJiraApiRequester.cs" />
     <Compile Include="Jira\IJiraApiRequestFactory.cs" />
     <Compile Include="Jira\JiraApiRequester.cs" />
     <Compile Include="Jira\JiraApiRequestFactory.cs" />
-    <Compile Include="NativeMethods.cs" />
+    <Compile Include="Helpers\NativeMethods.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -78,60 +78,60 @@
     <Compile Include="RestSharp\RestClientFactory.cs" />
     <Compile Include="RestSharp\IRestClientFactory.cs" />
     <Compile Include="RestSharp\IRestRequestFactory.cs" />
-    <Compile Include="PersistedIssue.cs" />
-    <Compile Include="JiraApi\SearchResult.cs" />
-    <Compile Include="JiraApi\Filter.cs" />
-    <Compile Include="Settings.cs" />
-    <Compile Include="WorklogForm.cs">
+    <Compile Include="Settings\PersistedIssue.cs" />
+    <Compile Include="Jira\DTO\SearchResult.cs" />
+    <Compile Include="Jira\DTO\Filter.cs" />
+    <Compile Include="Settings\Settings.cs" />
+    <Compile Include="UI\WorklogForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="WorklogForm.Designer.cs">
+    <Compile Include="UI\WorklogForm.Designer.cs">
       <DependentUpon>WorklogForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="JiraApi\Issue.cs" />
-    <Compile Include="IssueControl.cs">
+    <Compile Include="Jira\DTO\Issue.cs" />
+    <Compile Include="UI\IssueControl.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="JiraApi\IssueFields.cs" />
+    <Compile Include="Jira\DTO\IssueFields.cs" />
     <Compile Include="Jira\JiraClient.cs" />
-    <Compile Include="JiraTimeHelpers.cs" />
-    <Compile Include="LoginForm.cs">
+    <Compile Include="Helpers\JiraTimeHelpers.cs" />
+    <Compile Include="UI\LoginForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="LoginForm.Designer.cs">
+    <Compile Include="UI\LoginForm.Designer.cs">
       <DependentUpon>LoginForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="MainForm.cs">
+    <Compile Include="UI\MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="MainForm.Designer.cs">
+    <Compile Include="UI\MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SettingsForm.cs">
+    <Compile Include="UI\SettingsForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="SettingsForm.Designer.cs">
+    <Compile Include="UI\SettingsForm.Designer.cs">
       <DependentUpon>SettingsForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="WatchTimer.cs" />
-    <EmbeddedResource Include="AboutForm.resx">
+    <Compile Include="Model\WatchTimer.cs" />
+    <EmbeddedResource Include="UI\AboutForm.resx">
       <DependentUpon>AboutForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="EditTimeForm.resx">
+    <EmbeddedResource Include="UI\EditTimeForm.resx">
       <DependentUpon>EditTimeForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="WorklogForm.resx">
+    <EmbeddedResource Include="UI\WorklogForm.resx">
       <DependentUpon>WorklogForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="IssueControl.resx">
+    <EmbeddedResource Include="UI\IssueControl.resx">
       <DependentUpon>IssueControl.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="LoginForm.resx">
+    <EmbeddedResource Include="UI\LoginForm.resx">
       <DependentUpon>LoginForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="MainForm.resx">
+    <EmbeddedResource Include="UI\MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -139,7 +139,7 @@
       <SubType>Designer</SubType>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
-    <EmbeddedResource Include="SettingsForm.resx">
+    <EmbeddedResource Include="UI\SettingsForm.resx">
       <DependentUpon>SettingsForm.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="packages.config" />
