+++ source/StopWatch/UI/AboutForm.cs
