+++ source/StopWatch/UI/AboutForm.resx
