+++ source/StopWatch/UI/EditTimeForm.cs
