+++ source/StopWatch/UI/EditTimeForm.resx
