+++ source/StopWatch/UI/IssueControl.cs
