+++ source/StopWatch/UI/IssueControl.resx
