+++ source/StopWatch/UI/LoginForm.cs
