+++ source/StopWatch/UI/LoginForm.resx
