+++ source/StopWatch/UI/SettingsForm.resx
